+++ ndk/src/main/jni/utils/stack_unwinder.c
@@ -1,6 +1,7 @@
 #include "stack_unwinder.h"
 #include "stack_unwinder_libcorkscrew.h"
 #include "stack_unwinder_libunwind.h"
+#include "stack_unwinder_libunwindstack.h"
 #include "stack_unwinder_simple.h"
 #include "string.h"
 #include <asm/siginfo.h>
@@ -9,18 +10,35 @@
 #include <ucontext.h>
 
 #define BSG_LIBUNWIND_LEVEL 21
+#define BSG_LIBUNWINDSTACK_LEVEL 15
+#define BSG_LIBUNWIND_LEVEL_ARM32 16
 #define BSG_LIBCORKSCREW_MIN_LEVEL 16
 #define BSG_LIBCORKSCREW_MAX_LEVEL 19
 
-bsg_unwinder bsg_get_unwind_type(int apiLevel, bool is32bit) {
-  if (apiLevel >= BSG_LIBUNWIND_LEVEL && bsg_configure_libunwind(is32bit)) {
-    return BSG_LIBUNWIND;
-  } else if (apiLevel <= BSG_LIBCORKSCREW_MAX_LEVEL &&
-             apiLevel >= BSG_LIBCORKSCREW_MIN_LEVEL &&
-             bsg_configure_libcorkscrew()) {
-    return BSG_LIBCORKSCREW;
+void bsg_set_unwind_types(int apiLevel, bool is32bit, bsg_unwinder *signal_type,
+                          bsg_unwinder *other_type) {
+#if defined(__arm__)
+  if (apiLevel >= BSG_LIBUNWIND_LEVEL_ARM32 && is32bit &&
+      bsg_configure_libunwind(is32bit)) {
+    if (apiLevel >= BSG_LIBUNWIND_LEVEL) {
+      *signal_type = BSG_LIBUNWIND;
+    } else if (bsg_configure_libcorkscrew()) {
+      *signal_type = BSG_LIBCORKSCREW;
+    } else {
+      *signal_type = BSG_CUSTOM_UNWIND;
+    }
+    *other_type = BSG_LIBUNWIND;
+    return;
   }
-  return BSG_CUSTOM_UNWIND;
+#endif
+  if (apiLevel >= BSG_LIBUNWINDSTACK_LEVEL) {
+    bsg_configure_libunwind(is32bit);
+    *signal_type = BSG_LIBUNWINDSTACK;
+    *other_type = BSG_LIBUNWIND;
+  } else {
+    *signal_type = BSG_CUSTOM_UNWIND;
+    *other_type = BSG_CUSTOM_UNWIND;
+  }
 }
 
 void bsg_insert_fileinfo(ssize_t frame_count,
@@ -46,7 +64,10 @@ ssize_t bsg_unwind_stack(bsg_unwinder unwind_style,
                          bsg_stackframe stacktrace[BUGSNAG_FRAMES_MAX],
                          siginfo_t *info, void *user_context) {
   ssize_t frame_count = 0;
-  if (unwind_style == BSG_LIBUNWIND) {
+  if (unwind_style == BSG_LIBUNWINDSTACK) {
+    frame_count =
+        bsg_unwind_stack_libunwindstack(stacktrace, info, user_context);
+  } else if (unwind_style == BSG_LIBUNWIND) {
     frame_count = bsg_unwind_stack_libunwind(stacktrace, info, user_context);
   } else if (unwind_style == BSG_LIBCORKSCREW) {
     frame_count = bsg_unwind_stack_libcorkscrew(stacktrace, info, user_context);