+++ sdk/src/androidTest/java/com/bugsnag/android/AppDataTest.java
@@ -44,7 +44,7 @@ public class AppDataTest {
 
     @After
     public void tearDown() throws Exception {
-        Async.cancelTasks();
+        client.getOrientationListener().disable();
     }
 
     @Test