+++ sdk/src/androidTest/java/com/bugsnag/android/ClientConfigTest.java
@@ -35,7 +35,7 @@ public class ClientConfigTest {
 
     @After
     public void tearDown() throws Exception {
-        Async.cancelTasks();
+        client.getOrientationListener().disable();
     }
 
     @Test