+++ sdk/src/androidTest/java/com/bugsnag/android/ClientNotifyTest.java
@@ -36,7 +36,7 @@ public class ClientNotifyTest {
 
     @After
     public void tearDown() throws Exception {
-        Async.cancelTasks();
+        client.getOrientationListener().disable();
     }
 
     @Test